backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
pandas
google-genai
python-dotenv